app/requirements.txt
azure-functions
azure-storage-blob
azure-search-documents
openai
pymupdf
langchain-text-splitters
pydantic-settings
fastapi
uvicorn
langchain-core
langchain-openai
aiohttp
python-dotenv
azure-cosmos